requirements.txt
anyio==3.6.2

argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0

beautifulsoup4==4.11.1
bleach==6.0.0
blis==0.7.9
catalogue==1.0.2
certifi==2022.12.7
chardet==5.0.0
cinemagoer==2022.12.27
click==8.1.3
cymem==2.0.7
defusedxml==0.7.1
Django==4.1.3
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
docx2txt==0.8
exceptiongroup==1.0.4
fastjsonschema==2.16.2
fqdn==1.5.1
h11==0.14.0
html5lib==1.
IMDbPY==2022.7.9
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.1
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
langcodes==3.3.0
lxml==4.9.1
MarkupSafe==2.1.1
mistune==2.0.4
murmurhash==1.0.9
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
outcome==1.2.0
pandas==1.5.2
pandocfilters==1.5.0
pathy==0.10.0
pdfminer.six==20221105
plac==1.1.3
preshed==3.0.8
prometheus-client==0.16.0
pyadl==0.1
pycryptodome==3.16.0
pydantic==1.10.2
pyparsing==3.0.9
pyresparser==1.0.6
pyrsistent==0.19.2
python-json-logger==2.0.4
pytz==2022.6
PyYAML==6.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31

rfc3339-validator==0.1.4
rfc3986-validator==0.1.1

scikit-learn==1.1.3
scikit-surprise==1.1.3
scipy==1.9.3
selenium==4.6.1
Send2Trash==1.8.0
sklearn==0.0.post1
smart-open==5.2.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==2.3.8
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.46
sqlparse==0.4.3
srsly==1.0.6
terminado==0.17.1
thinc==7.4.6
threadpoolctl==3.1.0
tinycss2==1.2.1
trio==0.22.0
trio-websocket==0.9.2
typer==0.7.0
uri-template==1.2.0
wasabi==0.10.1
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
whitenoise==6.2.0
widgetsnbextension==4.0.5
WMI==1.4.9
wsproto==1.2.0
